refactor(siu): mv error, utils into siu; tests passing
BREAKING CHANGE: siuba.error and siuba.utils changed to siuba.siu.error, etc..
